fix(arm): improve ARM build process and compatibility
spdk-rs used to not build on arm solely due to the
nix environment setup. There were issues with compiler
flags (-msse4 is x86 only) and some base environment
configuration. By fixing the environment we were able
to get a successful build of spdk-rs.

On top of these fixes we are able to test them in the
github action runner which runs nix-shell. This in turn
compiles spdk from source (openebs/spdk).

Signed-off-by: Max Newcomer <maxwnewcomer@gmail.com>